+++ source/common/config/new_grpc_mux_impl.h
@@ -24,7 +24,7 @@ class NewGrpcMuxImpl : public GrpcMux,
                        public GrpcStreamCallbacks<envoy::api::v2::DeltaDiscoveryResponse>,
                        Logger::Loggable<Logger::Id::config> {
 public:
-  NewGrpcMuxImpl(Grpc::RawAsyncClientPtr async_client, Event::Dispatcher& dispatcher,
+  NewGrpcMuxImpl(Grpc::RawAsyncClientPtr&& async_client, Event::Dispatcher& dispatcher,
                  const Protobuf::MethodDescriptor& service_method, Runtime::RandomGenerator& random,
                  Stats::Scope& scope, const RateLimitSettings& rate_limit_settings,
                  const LocalInfo::LocalInfo& local_info);